radiofeed/podcasts/tests/test_tasks.py
from radiofeed.podcasts.tasks import parse_podcast_feed, send_recommendations_email


class TestTasks:
    def test_parse_podcast_feed(self, podcast, mocker):
        patched = mocker.patch(
            "radiofeed.podcasts.tasks.feed_parser.parse_podcast_feed"
        )
        parse_podcast_feed(podcast.id)
        patched.assert_called_with(podcast)

    def test_parse_podcast_feed_podcast_not_found(self, db, mocker):
        patched = mocker.patch(
            "radiofeed.podcasts.tasks.feed_parser.parse_podcast_feed"
        )
        parse_podcast_feed(1234)
        patched.assert_not_called()

    def test_send_recommendations_email(self, mocker, user):
        patched = mocker.patch(
            "radiofeed.podcasts.tasks.emails.send_recommendations_email"
        )
        send_recommendations_email(user.id)
        patched.assert_called_with(user)

    def test_send_recommendations_email_user_not_found(self, db, mocker):
        patched = mocker.patch(
            "radiofeed.podcasts.tasks.emails.send_recommendations_email"
        )
        send_recommendations_email(1234)
        patched.assert_not_called()
